backend/requirements.txt
# ── Django & REST ─────────────────────────────────────────────────────────────
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-celery-results==2.5.1

# ── Database ──────────────────────────────────────────────────────────────────
mysqlclient==2.2.3          # MySQL driver (swap to psycopg2 for Postgres)

# ── Async / Queue ─────────────────────────────────────────────────────────────
celery==5.3.6
redis==5.0.1

# ── AI / LLM ──────────────────────────────────────────────────────────────────
openai==1.12.0              # Works with OpenAI API AND LM Studio (same client)

# ── Vector DB ─────────────────────────────────────────────────────────────────
chromadb==0.4.22

# ── Scraping ──────────────────────────────────────────────────────────────────
selenium==4.17.2
beautifulsoup4==4.12.3
webdriver-manager==4.0.1    # Auto-downloads chromedriver

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
gunicorn==21.2.0            # Production WSGI server
whitenoise==6.6.0           # Static file serving
